Update UKCP18_frequency.json
STASH output from the underlying models is available on time profiles of both 3 hourly and 6 hourly means (with underlying data sampled every timestep) and as instantaneous snapshots (at 00Z, 03Z, 06Z etc.)

It is unclear to me exactly what the original CMIP frequency profiles of "3hr" and "6hr" are, since they are described as "sampled every 3/6 hours" which to me suggests instantaneous snapshots.  So I've added in 3hrMean and 6hrMean.  If my interpretation of the CMIP fields is wrong, and they are instead means rather than instantaneous snapshots, my additions should be changed to 3hrInst and 6hrInst.